refactor_payment_agent_appstore
